+++ src/cpp/inc/av/PAvStreaming.h
@@ -25,7 +25,7 @@ struct PAvStreaming {
     //
     // 流的长度，对于编码的时候，0表示不确定视频时长
     //
-    int timeDuration;
+    long long timeDuration;
 
     //
     // 样本描述类型
